\name{dat_gene2path_RNA}
\alias{dat_gene2path_RNA}
\docType{data}
\title{
demo result of gene2pathway function from RNA seq data
}
\description{
demo result of gene2pathway function from RNA seq data
}
\usage{data("dat_gene2path_RNA")}
\format{
  A list with 2 elements.
  \describe{
    \item{\code{gene2pathway_result.2}}{a data frame of gene2pathway test result}
    \item{\code{gene2pathway_result.FET}}{a data frame of fisher's test result}
  }
}

\value{
A list, with two elements. one element is a data frame about "gene2pathway_result.2"",
another data frame is about "gene2pathway_result.FET".
}

\keyword{datasets}
