\name{Des_BP_list}
\alias{Des_BP_list}
\docType{data}
\title{
Description of GO term BP
}
\description{
Description of GO term BP
}
\usage{data("Des_BP_list")}
\format{
A list, list names are Go term BP's IDs, list elements are GO term BP's descritpion respectively.
}
\value{
A list, list names are Go term BP's IDs, list elements are GO term BP's descritpion respectively.
}
\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}

\examples{
data(Des_BP_list)
head(names(Des_BP_list))
head(Des_BP_list)

}
\keyword{datasets}
