\name{Des_CC_list}
\alias{Des_CC_list}
\docType{data}
\title{
Description of GO term CC
}
\description{
Description of GO term CC
}
\usage{data("Des_CC_list")}
\format{
A list, list names are Go term CC's IDs, list elements are GO term CC's descritpion respectively.
}
\value{
A list, list names are Go term CC's IDs, list elements are GO term CC's descritpion respectively.
}
\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}

\examples{
data(Des_CC_list)
head(names(Des_CC_list))
head(Des_CC_list)

}
\keyword{datasets}
