\name{Des_MF_list}
\alias{Des_MF_list}
\docType{data}
\title{
Description of GO term MF
}
\description{
Description of GO term MF
}
\usage{data("Des_MF_list")}
\format{
A list, list names are Go term MF's IDs, list elements are GO term MF's descritpion respectively.
}
\value{
A list, list names are Go term MF's IDs, list elements are GO term MF's descritpion respectively.
}
\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}

\examples{
data(Des_MF_list)
head(names(Des_MF_list))
head(Des_MF_list)

}
\keyword{datasets}
