\name{gencode_coding}
\alias{gencode_coding}
\docType{data}
\title{
Coding gene list in GENCODE 20.
}
\description{
Coding gene list in GENCODE 20.
}
\usage{data("gencode_coding")}
\format{
A vector including all coding gene symbols in GENCODE version20.
}
\value{
A character vector including all coding gene symbols in GENCODE version20.
}

\source{
GENCODE 20
}
\references{
Harrow J, et al. (2012) GENCODE: The reference human genome annotation for The ENCODE Project
}
\examples{
data(gencode_coding)
head(gencode_coding)
}
\keyword{datasets}
