\name{gene_description}
\alias{gene_description}
\docType{data}
\title{
demo data of adding gene description
}
\description{
demo data of adding gene description
}
\usage{data("gene_description")}
\format{
  A data frame with 561 observations on the following 2 variables.
  \describe{
    \item{\code{hgnc_symbol}}{a characteristic vector}
    \item{\code{description}}{a characteristic vector}
  }
}

\value{
A characteristic matrix of gene symbols and descriptions.
}

\keyword{datasets}
