\name{GO_BP_list}
\alias{GO_BP_list}
\docType{data}
\title{
Gene symbol list of GO term BP
}
\description{
Gene symbol list of GO term BP
}
\usage{data("GO_BP_list")}
\format{
A list, list names are Go term BP's IDs, list elements are gene symbols in GO term BP.
}
\value{
A list, list names are Go term BP's IDs, list elements are gene symbols in GO term BP.
}

\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}

\examples{
data(GO_BP_list)
head(GO_BP_list)

}
\keyword{datasets}
