\name{GO_CC_list}
\alias{GO_CC_list}
\docType{data}
\title{
Gene symbol list of GO term CC
}
\description{
Gene symbol list of GO term CC
}
\usage{data("GO_CC_list")}
\format{
A list, list names are Go term CC's IDs, list elements are gene symbols in GO term CC.
}
\value{
A list, list names are Go term CC's IDs, list elements are gene symbols in GO term CC.
}
\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}

\examples{
data(GO_CC_list)
head(GO_CC_list)

}
\keyword{datasets}
