\name{GO_GENCODE_df_mm_vM1}
\alias{GO_GENCODE_df_mm_vM1}
\docType{data}
\title{
intersection genes of GO terms and GENCODE vM1
}
\description{
intersection genes of GO terms and GENCODE vM1
}
\usage{data("GO_GENCODE_df_mm_vM1")}
\format{
  A data frame.
}
\value{
  A data frame.
}
\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}
\references{
Harrow J, et al. (2012) GENCODE: The reference human genome annotation for The ENCODE Project
}
\examples{
data(GO_GENCODE_df_mm_vM1)
head(GO_GENCODE_df_mm_vM1)
}
\keyword{datasets}
