\name{GO_MF_list}
\alias{GO_MF_list}
\docType{data}
\title{
Gene symbol list of GO term MF
}
\description{
Gene symbol list of GO term MF
}
\usage{data("GO_MF_list")}
\format{
A list, list names are Go term MF's IDs, list elements are gene symbols in GO term MF.
}
\value{
A list, list names are Go term MF's IDs, list elements are gene symbols in GO term MF.
}
\source{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.0.0.
}

\examples{
data(GO_MF_list)
head(GO_MF_list)

}
\keyword{datasets}
