\name{Msig_GENCODE_df_hg_v20}
\alias{Msig_GENCODE_df_hg_v20}
\docType{data}
\title{
intersection genes of MSigDB collection and GENCODE 20
}
\description{
intersection genes of MSigDB collection and GENCODE 20
}
\usage{data("Msig_GENCODE_df_hg_v20")}
\format{
  A data frame.

}
\value{
  A data frame.
}
\references{
Harrow J, et al. (2012) GENCODE: The reference human genome annotation for The ENCODE Project.
Subramanian, Tamayo, et al. (2005, PNAS 102, 15545-15550).
}
\examples{
data(Msig_GENCODE_df_hg_v20)
head(Msig_GENCODE_df_hg_v20)
}
\keyword{datasets}
