\name{Msig_GENCODE_df_hg_v36}
\alias{Msig_GENCODE_df_hg_v36}
\docType{data}
\title{
intersection genes of MSigDB collection and GENCODE 36
}
\description{
intersection genes of MSigDB collection and GENCODE 36
}
\usage{data("Msig_GENCODE_df_hg_v36")}
\format{
  A data frame.

}
\value{
  A data frame.
}
\references{
Harrow J, et al. (2012) GENCODE: The reference human genome annotation for The ENCODE Project.
Subramanian, Tamayo, et al. (2005, PNAS 102, 15545-15550).
}
\examples{
data(Msig_GENECODE_df_hg_v36)
head(Msig_GENECODE_df)
}
\keyword{datasets}
