\name{MsigDB_C5}
\alias{MsigDB_C5}
\docType{data}
\title{
MsigDB data set
}
\description{
.gsa format data
}
\usage{data("MsigDB_C5")}
\format{
  A list with 3 sub lists.
  \describe{
    \item{\code{genesets}}{a list of gene names}
    \item{\code{geneset.names}}{a list of geneset names}
    \item{\code{geneset.descriptions}}{a list of geneset descriptions}
  }
}

\value{
A "GSA.genesets" object from GSA package.
}
\keyword{datasets}
